requirements.txt
requests==2.20.0
aiohttp==3.7.4
aiofiles==0.5.0
python-dateutil==2.6.1
pillow==8.2.0
motor==1.2
pymongo==3.7.1
emoji